+++ test/integration/shpec/run_exit_status_preserved.sh
@@ -1,16 +1,18 @@
-describe "ide preserves exit status"
+describe "ide command: run preserves exit status"
   # make absolute path out of relative
   IDE_PATH=$(readlink -f "./ide")
 
   describe 'when IDE_DRIVER="docker"'
     describe 'when exit 164 in docker container'
       it "returns 164"
-        # clean up before test, if there is no such image, docker will
-        # return error, ignore that
-        docker rmi alpine:3.4
         message="$(cd test/docker/dummyide-usage && IDE_LOG_LEVEL=debug ${IDE_PATH} 'echo abc && exit 164')"
-        assert equal "$?" "164"
-        assert do_match "$message" "abc"
+        exit_status="$?"
+        it "exits with status 164"
+          assert equal "$exit_status" "164"
+        end
+        it "docker run command returns 'abc'"
+          assert do_match "$message" "abc"
+        end
       end
     end
   end
@@ -18,8 +20,13 @@ describe "ide preserves exit status"
     describe 'when exit 164 in docker container'
       it "returns 164"
         message="$(cd test/docker-compose/default && IDE_LOG_LEVEL=debug ${IDE_PATH} --force_not_interactive 'echo abc && exit 164')"
-        assert equal "$?" "164"
-        assert do_match "$message" "abc"
+        exit_status="$?"
+        it "exits with status 164"
+          assert equal "$exit_status" "164"
+        end
+        it "docker-compose run command returns 'abc'"
+          assert do_match "$message" "abc"
+        end
       end
     end
   end