+++ test/unit/shpec/command.sh
@@ -1,25 +1,28 @@
 describe "ide command"
   # make absolute path out of relative
   IDE_PATH=$(readlink -f "./ide")
+  # Tests that IDE can be run with: --command or -c option or
+  # without any command option (using default command).
+  # Warn: this uses the main Idefile.
 
-  describe "when run without --command option"
+  describe "when run without --command or --c option"
     message="$(IDE_LOG_LEVEL=debug ${IDE_PATH} --dryrun)"
     exit_status="$?"
     it "exits with status 0"
       assert equal "$exit_status" "0"
     end
-    it "informs that ide command is run"
+    it "informs that ide command is: run"
       assert do_match "$message" "ide_command: run"
     end
   end
-  describe "when run with --command option"
+  describe "when run explicitly with default command"
     describe "when run with '--command run'"
       message="$(IDE_LOG_LEVEL=debug ${IDE_PATH} --command run --dryrun)"
       exit_status="$?"
       it "exits with status 0"
         assert equal "$exit_status" "0"
       end
-      it "informs that ide command is run"
+      it "informs that ide command is: run"
         assert do_match "$message" "ide_command: run"
       end
     end
@@ -29,17 +32,19 @@ describe "ide command"
       it "exits with status 0"
         assert equal "$exit_status" "0"
       end
-      it "informs that ide command is run"
+      it "informs that ide command is: run"
         assert do_match "$message" "ide_command: run"
       end
     end
+  end
+  describe "when run with some valid, not default command"
     describe "when run with '--command pull'"
       message="$(IDE_LOG_LEVEL=debug ${IDE_PATH} --command pull --dryrun)"
       exit_status="$?"
       it "exits with status 0"
         assert equal "$exit_status" "0"
       end
-      it "informs that ide command is pull"
+      it "informs that ide command is: pull"
         assert do_match "$message" "ide_command: pull"
       end
     end
@@ -49,10 +54,12 @@ describe "ide command"
       it "exits with status 0"
         assert equal "$exit_status" "0"
       end
-      it "informs that ide command is pull"
+      it "informs that ide command is: pull"
         assert do_match "$message" "ide_command: pull"
       end
     end
+  end
+  describe "when run with some invalid command"
     describe "when run with '--command bla'"
       message="$(IDE_LOG_LEVEL=debug ${IDE_PATH} --command bla --dryrun)"
       exit_status="$?"